api/requirements.txt
Django==4.2.4
psycopg2>=2.9.6
djangorestframework>=3.14.0
Markdown==3.4.3
django-filter==23.2
Pillow==10.0.0